requirements.txt
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-openai
llama-index-llms-anthropic
chromadb
python-docx
anthropic
openai
python-dotenv
pydantic
rich
pytest
